im going to cry

+++ nmapscan.py
@@ -2,8 +2,12 @@ import nmap
 import os
 
 Scan = nmap.PortScanner()
-os.remove("oldOutput.txt")
-os.rename("newOutput.txt", "oldOutput.txt")
+
+if os.path.exists("oldOutput.txt"):
+    os.remove("oldOutput.txt")
+
+if os.path.exists("newOutput.txt"):
+    os.rename("newOutput.txt", "oldOutput.txt")
 
 hostFile = open('targets.txt')
 ipFile = open("IP list.txt", "w")
@@ -28,22 +32,23 @@ if len(targets) != 0:
         print('----------------------------------------------------')
         print('Host : %s (%s)' % (host, Scan[host].hostname()))
 
-        output.write(host)
+        output.write(host + ',')
 
         mac = "-"
         vendorName = "-"
 
         if 'mac' in Scan[host]['addresses']:
             mac = Scan[host]['addresses']['mac']
-            output.write(", " + mac)
+            output.write(mac)
+
+        output.write(',')
 
         print('State : %s' % Scan[host].state())
         if Scan[host].all_protocols():
-            output.write(", ")
             for proto in Scan[host].all_protocols():
                 print('----------')
                 print('Protocol : %s' % proto)
-                output.write(proto + "; ")
+                #output.write(proto + "; ")
                 ports = Scan[host][proto].keys()
 
                 fst = True
@@ -51,8 +56,8 @@ if len(targets) != 0:
                     if fst:
                         fst = False
                     else:
-                        output.write(", ")
-                    output.write(str(port) + ": " + Scan[host][proto][port]['state'])
+                        output.write(",")
+                    output.write(proto.upper() + str(port) + ":" + Scan[host][proto][port]['state'])
                     print('port : %s\t state : %s' % (port, Scan[host][proto][port]['state']))
                 output.write("\n")
         else:
